+++ src/debug_terminal/arguments.rs
@@ -232,7 +232,7 @@ impl StateArgs {
     }
 
     pub fn from_state(
-        current_state: &DVector<Complex<f64>>,
+        current_state: &DVector<Complex<f32>>,
         state_args: &StateArgs,
     ) -> Result<Vec<IndexedState>, StateError> {
         let state_size = current_state.len() - 1;
