+++ src/gate.rs
@@ -1,5 +1,5 @@
 use std::{
-    f64::consts::{FRAC_1_SQRT_2, PI},
+    f32::consts::{FRAC_1_SQRT_2, PI},
     ops::{BitAnd, Shl, Shr, ShrAssign},
 };
 
@@ -95,7 +95,7 @@ pub enum GateType {
     Z,
     H,
     SWAP,
-    U(f64, f64, f64),
+    U(f32, f32, f32),
     S,
 }
 
@@ -131,31 +131,31 @@ pub struct Gate {
 
 impl Gate {
     #[rustfmt::skip]
-    pub const PAULI_X_DATA: [Complex<f64>; 4] = [
+    pub const PAULI_X_DATA: [Complex<f32>; 4] = [
         cart!(0.0), cart!(1.0),
         cart!(1.0), cart!(0.0),
     ];
 
     #[rustfmt::skip]
-    pub const PAULI_Y_DATA: [Complex<f64>; 4] = [
+    pub const PAULI_Y_DATA: [Complex<f32>; 4] = [
         cart!(0.0), cart!(0.0, -1.0),
         cart!(0.0, 1.0), cart!(0.0),
     ];
 
     #[rustfmt::skip]
-    pub const PAULI_Z_DATA: [Complex<f64>; 4] = [
+    pub const PAULI_Z_DATA: [Complex<f32>; 4] = [
         cart!(1.0), cart!(0.0),
         cart!(0.0), cart!(-1.0, 0.0),
     ];
 
     #[rustfmt::skip]
-    pub const HADAMARD_DATA: [Complex<f64>; 4] = [
+    pub const HADAMARD_DATA: [Complex<f32>; 4] = [
         cart!(FRAC_1_SQRT_2, 0.0), cart!(FRAC_1_SQRT_2, 0.0),
         cart!(FRAC_1_SQRT_2, 0.0), cart!(-FRAC_1_SQRT_2, 0.0),
     ];
 
     #[rustfmt::skip]
-    pub const SWAP_DATA: [Complex<f64>; 16] = [
+    pub const SWAP_DATA: [Complex<f32>; 16] = [
         cart!(1.0), cart!(0.0), cart!(0.0), cart!(0.0),
         cart!(0.0), cart!(0.0), cart!(1.0), cart!(0.0),
         cart!(0.0), cart!(1.0), cart!(0.0), cart!(0.0),
@@ -163,7 +163,7 @@ impl Gate {
     ];
 
     #[rustfmt::skip]
-    pub const PHASE_S_DATA: [Complex<f64>; 4] = [
+    pub const PHASE_S_DATA: [Complex<f32>; 4] = [
         cart!(1.0), cart!(0.0),
         cart!(0.0), cart!(0.0, 1.0),
     ];
